Edit info for online manga + Custom covers update
Yes you read that right. It's back! Oh god it's back

Instead of modifying the db, an external json file is made holding the custom info for your library (meaning it's even easier to remove should I so choose)
Reworking to just override the variable and use said var instead of having the current/original logic that existed before

Custom covers are now saved in a new folder, likewise to upstream
Also like upstream, custom covers can be added to manga without covers (closes #49)

(I'm so sorry Carlos)

(cherry picked from commit d3ec230d4baa8584118dc30807728305715db25b)

+++ app/src/main/java/eu/kanade/tachiyomi/data/cache/CoverCache.kt
@@ -6,6 +6,7 @@ import coil.Coil
 import eu.kanade.tachiyomi.R
 import eu.kanade.tachiyomi.data.database.DatabaseHelper
 import eu.kanade.tachiyomi.data.database.models.Manga
+import eu.kanade.tachiyomi.data.database.models.MangaImpl
 import eu.kanade.tachiyomi.util.storage.DiskUtil
 import eu.kanade.tachiyomi.util.system.executeOnIO
 import eu.kanade.tachiyomi.util.system.toast
@@ -29,11 +30,16 @@ import java.io.InputStream
  */
 class CoverCache(val context: Context) {
 
-    /**
-     * Cache directory used for cache management.
-     */
-    private val cacheDir = context.getExternalFilesDir("covers")
-        ?: File(context.filesDir, "covers").also { it.mkdirs() }
+    companion object {
+        private const val COVERS_DIR = "covers"
+        private const val CUSTOM_COVERS_DIR = "covers/custom"
+    }
+
+    /** Cache directory used for cache management.*/
+    private val cacheDir = getCacheDir(COVERS_DIR)
+
+    /** Cache directory used for custom cover cache management.*/
+    private val customCoverCacheDir = getCacheDir(CUSTOM_COVERS_DIR)
 
     fun getChapterCacheSize(): String {
         return Formatter.formatFileSize(context, DiskUtil.getDirectorySize(cacheDir))
@@ -50,7 +56,7 @@ class CoverCache(val context: Context) {
             val files = cacheDir.listFiles()?.iterator() ?: return@launch
             while (files.hasNext()) {
                 val file = files.next()
-                if (file.name !in urls) {
+                if (file.isFile && file.name !in urls) {
                     deletedSize += file.length()
                     file.delete()
                 }
@@ -66,6 +72,45 @@ class CoverCache(val context: Context) {
     }
 
     /**
+     * Returns the custom cover from cache.
+     *
+     * @param manga the manga.
+     * @return cover image.
+     */
+    fun getCustomCoverFile(manga: Manga): File {
+        return File(customCoverCacheDir, DiskUtil.hashKeyForDisk(manga.id.toString()))
+    }
+
+    /**
+     * Saves the given stream as the manga's custom cover to cache.
+     *
+     * @param manga the manga.
+     * @param inputStream the stream to copy.
+     * @throws IOException if there's any error.
+     */
+    @Throws(IOException::class)
+    fun setCustomCoverToCache(manga: Manga, inputStream: InputStream) {
+        getCustomCoverFile(manga).outputStream().use {
+            inputStream.copyTo(it)
+            Coil.imageLoader(context).invalidate(manga.key())
+        }
+    }
+
+    /**
+     * Delete custom cover of the manga from the cache
+     *
+     * @param manga the manga.
+     * @return whether the cover was deleted.
+     */
+    fun deleteCustomCover(manga: Manga): Boolean {
+        val result = getCustomCoverFile(manga).let {
+            it.exists() && it.delete()
+        }
+        Coil.imageLoader(context).invalidate(manga.key())
+        return result
+    }
+
+    /**
      * Returns the cover from cache.
      *
      * @param thumbnailUrl the thumbnail url.
@@ -75,19 +120,11 @@ class CoverCache(val context: Context) {
         return File(cacheDir, manga.key())
     }
 
-    /**
-     * Copy the given stream to this cache.
-     *
-     * @param thumbnailUrl url of the thumbnail.
-     * @param inputStream the stream to copy.
-     * @throws IOException if there's any error.
-     */
-    @Throws(IOException::class)
-    fun copyToCache(manga: Manga, inputStream: InputStream) {
-        // Get destination file.
-        val destFile = getCoverFile(manga)
-
-        destFile.outputStream().use { inputStream.copyTo(it) }
+    fun deleteFromCache(name: String?) {
+        if (name.isNullOrEmpty()) return
+        val file = getCoverFile(MangaImpl().apply { thumbnail_url = name })
+        Coil.imageLoader(context).invalidate(file.name)
+        if (file.exists()) file.delete()
     }
 
     /**
@@ -96,13 +133,21 @@ class CoverCache(val context: Context) {
      * @param thumbnailUrl the thumbnail url.
      * @return status of deletion.
      */
-    fun deleteFromCache(manga: Manga, deleteMemoryCache: Boolean = true) {
+    fun deleteFromCache(
+        manga: Manga,
+        deleteCustom: Boolean = true
+    ) {
         // Check if url is empty.
         if (manga.thumbnail_url.isNullOrEmpty()) return
 
         // Remove file
         val file = getCoverFile(manga)
-        if (deleteMemoryCache) Coil.imageLoader(context).invalidate(file.name)
+        if (deleteCustom) deleteCustomCover(manga)
         if (file.exists()) file.delete()
+    }
+
+    private fun getCacheDir(dir: String): File {
+        return context.getExternalFilesDir(dir)
+            ?: File(context.filesDir, dir).also { it.mkdirs() }
     }
 }
